Add suppression filtering to HTML generation and fix CS0649 warning (#1233, #1234)
- Filter suppressed diagnostics in HtmlCodeWriter.WriteAllAsync and WriteAllDiffAsync
- Pass DiagnosticSuppressions to HTML diff generation in CompileTimeAspectPipeline
- Add CS0649 (field never assigned) suppression to IntroduceDependencyAttribute
- Rename HighlightingTestRunner to HtmlGenerationTestRunner and move to Metalama.Testing.AspectTesting
- Add HTML generation test for DependencyInjection to verify CS0649 suppression
