+++ Metalama.Extensions/src/tests/Metalama.Extensions.DependencyInjection.AspectTests/Html/EarlyRequired_Html.cs.html
@@ -1,0 +1,102 @@
+
+<html>
+    <head>
+        <style>
+            .cr-CompileTime,
+            .cr-Conflict,
+            .cr-TemplateKeyword,
+            .cr-Dynamic,
+            .cr-CompileTimeVariable,
+            .cr-GeneratedCode
+            {
+                background-color: rgba(50,50,90,0.1);
+            }
+
+            .cr-NeutralTrivia
+            {
+                background-color: rgba(0,255,0,0.1);
+            }
+
+            .cr-TemplateKeyword
+            {
+                color: rgb(250, 0, 250) !important;
+                font-weight: bold;
+            }
+
+            .cr-Dynamic
+            {
+                text-decoration: underline;
+            }
+
+            .cr-CompileTimeVariable
+            {
+                font-style: italic;
+            }
+
+            .diag-Warning
+            {
+                text-decoration: underline 1px wavy orange;
+            }
+
+            .diag-Error
+            {
+                text-decoration: underline 1px wavy red;
+            }
+
+         .diff-Imaginary {
+                display: block;
+                background-image: repeating-linear-gradient( -45deg, gray, gray 2px, transparent 2px, transparent 8px );
+            }
+
+
+            .legend
+            {
+                margin-top: 100px;
+            }
+        </style>
+    </head>
+    <body><pre><code class="nohighlight"><span class='line-number'>1</span><span class="cs-keyword">using</span> <span class="cs-namespace-name">Metalama</span><span class="cs-operator">.</span><span class="cs-namespace-name">Framework</span><span class="cs-operator">.</span><span class="cs-namespace-name">Aspects</span><span class="cs-punctuation">;</span>
+<span class='line-number'>2</span>
+<span class='line-number'>3</span><span class="cs-keyword">namespace</span> <span class="cs-namespace-name">Metalama</span><span class="cs-operator">.</span><span class="cs-namespace-name">Extensions</span><span class="cs-operator">.</span><span class="cs-namespace-name">DependencyInjection</span><span class="cs-operator">.</span><span class="cs-namespace-name">AspectTests</span><span class="cs-operator">.</span><span class="cs-namespace-name">Html</span><span class="cs-operator">.</span><span class="cs-namespace-name">EarlyRequired_Html</span><span class="cs-punctuation">;</span>
+<span class='line-number'>4</span>
+<span class='line-number' data-member='TargetClass'>5</span><span class="cs-comment">// This test verifies that CS0649/CS8618 suppressions from [IntroduceDependency] are correctly applied to input HTML.</span>
+<span class='line-number' data-member='TargetClass'>6</span><span class="cs-comment">// The [IntroduceDependency] attribute should suppress CS0649 (field never assigned) and CS8618 on the aspect field.</span>
+<span class='line-number' data-member='TargetClass'>7</span><span class="cs-comment">// Expected: CS0649 and CS8618 should NOT appear in the input HTML because the aspect suppresses them.</span>
+<span class='line-number' data-member='TargetClass'>8</span>
+<span class='line-number' data-member='TargetClass'>9</span><span class="cs-comment">// &lt;target&gt;</span>
+<span class='line-number' data-member='TargetClass'>10</span><span class="cs-keyword">public</span> <span class="cs-keyword">class</span> <span class="cs-class-name">TargetClass</span>
+<span class='line-number' data-member='TargetClass'>11</span><span class="cs-punctuation">{</span>
+<span class='line-number' data-member='TargetClass.Method'>12</span>    <span class="cs-punctuation">[</span><span class="cs-class-name">LogAspect</span><span class="cs-punctuation">]</span>
+<span class='line-number' data-member='TargetClass.Method'>13</span>    <span class="cs-keyword">public</span> <span class="cs-keyword">void</span> <span class="cs-method-name">Method</span><span class="cs-punctuation">()</span> <span class="cs-punctuation">{</span> <span class="cs-punctuation">}</span>
+<span class='line-number' data-member='TargetClass'>14</span><span class="cs-punctuation">}</span>
+<span class='line-number' data-member='TargetClass'>15</span>
+<span class='line-number' data-member='LogAspect'>16</span><span class="cs-keyword">public</span> <span class="cs-keyword">class</span> <span class="cs-class-name">LogAspect</span> <span class="cs-punctuation">:</span> <span class="cs-class-name">OverrideMethodAspect</span>
+<span class='line-number' data-member='LogAspect'>17</span><span class="cs-punctuation">{</span>
+<span class='line-number' data-member='LogAspect._formatProvider'>18</span>    <span class="cs-punctuation">[</span><span class="cs-class-name">IntroduceDependency</span><span class="cs-punctuation">]</span>
+<span class='line-number' data-member='LogAspect._formatProvider'>19</span>    <span class="cs-keyword">private</span> <span class="cs-keyword">readonly</span> <span class="cs-identifier">IFormatProvider</span> <span class="cs-field-name">_formatProvider</span><span class="cs-punctuation">;</span>
+<span class='line-number' data-member='LogAspect._formatProvider'>20</span>
+<span class='line-number' data-member='LogAspect.OverrideMethod'>21</span>    <span class="cs-keyword">public</span> <span class="cs-keyword">override</span> <span class="cs-keyword">dynamic</span><span class="cs-operator">?</span> <span class="cs-method-name">OverrideMethod</span><span class="cs-punctuation">()</span>
+<span class='line-number' data-member='LogAspect.OverrideMethod'>22</span>    <span class="cs-punctuation">{</span>
+<span class='line-number' data-member='LogAspect.OverrideMethod'>23</span>        <span class="cs-identifier">Console</span><span class="cs-operator">.</span><span class="cs-identifier">WriteLine</span><span class="cs-punctuation">(</span> <span class="cs-keyword">this</span><span class="cs-operator">.</span><span class="cs-field-name">_formatProvider</span> <span class="cs-punctuation">);</span>
+<span class='line-number' data-member='LogAspect.OverrideMethod'>24</span>
+<span class='line-number' data-member='LogAspect.OverrideMethod'>25</span>        <span class="cs-keyword cs-control">return</span> <span class="cs-class-name cs-static-symbol">meta</span><span class="cs-operator">.</span><span class="cs-method-name cs-static-symbol">Proceed</span><span class="cs-punctuation">();</span>
+<span class='line-number' data-member='LogAspect.OverrideMethod'>26</span>    <span class="cs-punctuation">}</span>
+<span class='line-number' data-member='LogAspect'>27</span><span class="cs-punctuation">}</span>
+<span class='line-number'>28</span>
+</code></pre>
+       <p class='legend'>Legend:</p>
+       <pre>
+<span class='cr-Default'>Default</span>
+<span class='cr-RunTime'>RunTime</span>
+<span class='cr-CompileTime'>CompileTime</span>
+<span class='cr-Dynamic'>Dynamic</span>
+<span class='cr-CompileTimeVariable'>CompileTimeVariable</span>
+<span class='cr-TemplateKeyword'>TemplateKeyword</span>
+<span class='cr-Conflict'>Conflict</span>
+<span class='cr-Excluded'>Excluded</span>
+<span class='cr-GeneratedCode'>GeneratedCode</span>
+<span class='cr-SourceCode'>SourceCode</span>
+<span class='cr-NeutralTrivia'>NeutralTrivia</span>
+       </pre>
+   </body>
+</html>
