+++ Metalama.Extensions/src/tests/Metalama.Extensions.DependencyInjection.AspectTests/Html/metalamaTests.json
@@ -1,5 +1,6 @@
 {
   "TestRunnerFactoryType": "Metalama.Testing.AspectTesting.HtmlGenerationTestRunnerFactory, Metalama.Testing.AspectTesting",
   "WriteInputHtml": true,
-  "WriteOutputHtml": true
+  "WriteOutputHtml": false,
+  "FormatOutput": true
 }
