--- Metalama.Framework/src/tests/Metalama.Framework.Tests.AspectTests/Runners/HighlightingTestRunnerFactory.cs
@@ -1,22 +1,0 @@
-// Copyright (c) 2020-2025 SharpCrafters s.r.o. and contributors.
-// SharpCrafters s.r.o. licenses this file to you under either the MIT license or a proprietary license, depending on the repository from which it was obtained.
-// Refer to LICENSE.md in the repository root for complete details.
-
-using JetBrains.Annotations;
-using Metalama.Framework.Engine.Services;
-using Metalama.Testing.AspectTesting;
-using Xunit.Abstractions;
-
-namespace Metalama.Framework.Tests.AspectTests.Runners
-{
-    [UsedImplicitly]
-    internal class HighlightingTestRunnerFactory : ITestRunnerFactory
-    {
-        public BaseTestRunner CreateTestRunner(
-            GlobalServiceProvider serviceProvider,
-            string? projectDirectory,
-            TestProjectReferences references,
-            ITestOutputHelper? logger )
-            => new HighlightingTestRunner( serviceProvider, projectDirectory, references, logger );
-    }
-}
