+++ Metalama.Framework/src/tests/Metalama.Framework.Tests.AspectTests/Tests/Formatting/metalamaTests.json
@@ -1,5 +1,5 @@
 {
-  "TestRunnerFactoryType": "Metalama.Framework.Tests.AspectTests.Runners.HighlightingTestRunnerFactory",
+  "TestRunnerFactoryType": "Metalama.Testing.AspectTesting.HtmlGenerationTestRunnerFactory, Metalama.Testing.AspectTesting",
   "WriteInputHtml": true,
   "FormatOutput": true,
   "CompareWhitespace":  true,
